Refactor controller package to api package that contains controllers, request and response objects per-feature
