+++ src/main/java/com/example/echo_api/api/v1/auth/request/SignUpReqest.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.persistence.dto.request;
+package com.example.echo_api.api.v1.auth.request;
 
 import com.example.echo_api.config.ValidationConfig;
